version with 1/2 tests

+++ src/main/java/ua/com/poseal/todo/services/UserService.java
@@ -1,34 +1,23 @@
 package ua.com.poseal.todo.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
 import ua.com.poseal.todo.domain.User;
-import ua.com.poseal.todo.exceptions.ResourceNotFoundException;
+import ua.com.poseal.todo.exceptions.UserNotFoundException;
+import ua.com.poseal.todo.exceptions.ResponseNotContentException;
 import ua.com.poseal.todo.exceptions.ResponseNotValidDataException;
 import ua.com.poseal.todo.repositories.UserRepository;
 
-import javax.annotation.PostConstruct;
 import javax.validation.ConstraintViolation;
 import javax.validation.Validator;
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import static ua.com.poseal.todo.TodoApplication.logger;
-
 @Service
 public class UserService {
 
-    private static final String USERS_CSV = "users_information.csv";
+    //    private static final String USERS_CSV = "users_information.csv";
     private final UserRepository repository;
     private final Validator validator;
 
@@ -41,14 +30,14 @@ public class UserService {
     public List<User> findAll() {
         List<User> users = repository.findAll();
         if (users.isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.NO_CONTENT, "Empty data");
+            throw new ResponseNotContentException("Empty data");
         }
         return users;
     }
 
     public User findById(Long id) {
         return repository.findById(id).orElseThrow(
-                () -> new ResourceNotFoundException(id));
+                () -> new UserNotFoundException(id));
     }
 
     public User createUser(User user) {
@@ -58,6 +47,16 @@ public class UserService {
 
     public User updateUser(Long id, User user) {
         validateUser(user);
+        user.setId(id);
+//        User userFromDB = repository.findById(id).orElseThrow(
+//                () -> new UserNotFoundException(id));
+//
+//        userFromDB.setFirstName(user.getFirstName());
+//        userFromDB.setLastName(user.getLastName());
+//        userFromDB.setIpn(user.getIpn());
+//
+//        return repository.save(user);
+
         return repository.findById(id)
                 .map(u -> {
                     u.setFirstName(user.getFirstName());
@@ -65,15 +64,14 @@ public class UserService {
                     u.setIpn(user.getIpn());
                     return repository.save(u);
                 })
-                .orElseGet(() -> repository.save(user));
+                .orElseThrow(() -> new UserNotFoundException(id));
     }
 
     public void deleteUser(Long id) {
-        try {
+        if (repository.findById(id).isPresent()) {
             repository.deleteById(id);
-        } catch (Exception ex) {
-            throw new ResourceNotFoundException(id);
         }
+        throw new UserNotFoundException(id);
     }
 
     private void validateUser(User user) {
@@ -89,25 +87,25 @@ public class UserService {
                 .collect(Collectors.joining(", "));
     }
 
-    @PostConstruct
-    private void initUsersFromCSV() {
-        List<User> usersList = new ArrayList<>();
-        try (InputStream resource = new ClassPathResource(USERS_CSV).getInputStream();
-             BufferedReader reader = new BufferedReader(new InputStreamReader(resource, StandardCharsets.UTF_8))) {
-            String line;
-            while ((line = reader.readLine()) != null) {
-                String[] usersString = line.split(";");
-                usersList.add(
-                        new User(Long.parseLong(
-                                usersString[0].trim()),
-                                usersString[1].trim(),
-                                usersString[2].trim(),
-                                usersString[3].trim())
-                );
-            }
-        } catch (IOException ex) {
-            logger.error("Error reading file {}", USERS_CSV, ex);
-        }
-        repository.saveAll(usersList);
-    }
+//    @PostConstruct
+//    private void initUsersFromCSV() {
+//        List<User> usersList = new ArrayList<>();
+//        try (InputStream resource = new ClassPathResource(USERS_CSV).getInputStream();
+//             BufferedReader reader = new BufferedReader(new InputStreamReader(resource, StandardCharsets.UTF_8))) {
+//            String line;
+//            while ((line = reader.readLine()) != null) {
+//                String[] usersString = line.split(";");
+//                usersList.add(
+//                        new User(Long.parseLong(
+//                                usersString[0].trim()),
+//                                usersString[1].trim(),
+//                                usersString[2].trim(),
+//                                usersString[3].trim())
+//                );
+//            }
+//        } catch (IOException ex) {
+//            logger.error("Error reading file {}", USERS_CSV, ex);
+//        }
+//        repository.saveAll(usersList);
+//    }
 }
